Ok So i converted it back to how it was although...
I've moved all of the state for fetchBuildWeeks up a level and passed
everything down. exact same problem. The problem is coming from that
function it-self to get the info to firebase is fine but the double
rendereing on ui is bad and immediatly goes away on refresh.  Have to
get this resolved. moving on for now